Making the client to be a class.
This simplifies things *a lot*.

I am considering also making the server a class.
But for now it is just a function.

I converted the client to a class, because I didn't know how to:
* Share state between a set of functions
* Expose some of the functions outside the module

The only ways I know to share state between functions are:
* Make the functions nested, with the toplevel function holding the
shared state (the functions' closures hold the state). However, this
effectively makes the top-level function a class with private methods.
* Use global variables. Use what now?

I think it looks cool, OOP + Functional. So, beat it.